chore(lib): move core functionality
